backend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
django-hosts===4.0
django-email-utils==1.0.0
emailutils==0.0.1
django-cors-headers==3.5.0
django-encrypted-filefield==0.2.2
django-polymorphic==3.0.0
libmagic==1.0
python-magic-bin==0.4.14
ffmpeg-python==0.2.0
